app/requirements.txt
Django
djangorestframework
psycopg2
requests